% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic1-package.R
\docType{package}
\name{arithmetic1}
\alias{arithmetic1}
\alias{arithmetic1-package}
\title{\code{arithmetic1} package}
\description{
An example package that does basic arithmetic.
}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
